requirements.txt
torch==1.11.0
torchvision==0.12.0
pandas==2.1.0
Pillow==10.0.0
scikit-learn==1.3.0